+++ packages/dataviews/src/components/dataviews-layouts/utils/test/use-selection-props.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getRangeSelection } from '../use-selection-gestures';
+import { getRangeSelection } from '../use-selection-props';
 
 const orderedIds = [ '1', '2', '3', '4', '5' ];